feature(endpoint): add record likes to gif endpoint [Start #212]
